Start on cell iterator
We're going to use a tagged union to represent the cell, so that
we can use an iterator to yield each cell from a page.

This commit just sets up the type for the tagged union, and the
conditional logic in the iter_next function for constructing
Cells correctly based on the page type.