Read values from records
Implements everything required for the basic .tables command.

At this point we can iterate through cells on a page, and for
each of those cells access the value for a column by it's index.

We still haven't needed to allocate any memory other than the
memory mapped file. All values are just  small struct on the stack
that either copies the data for integers and floats, or just
points back into the memory mapped data.